Extract PointCloudRenderable from PointCloud
